Add chunk reading of urlopen response.
Sometimes (esp. on slow connections for me, ie sat) the server will send
chunks of data without a zero-length end chunk. This triggers an
http.client.IncompleteRead exception in http.client.  This patch
accounts for that and loop-reads the chunks.

I do still see an occational what seems stream corruption still, usually
shown as the following sample:

  Unexpected error for command: list, error: Expecting property name
  enclosed in double quotes: line 1 column 44351 (char 44350)

This is from the json lib and is rare.

Signed-off-by: Alex Tsariounov <alext@manetu.com>