Change gql chunk read to retry on IncompleteRead
The issue with the previous attempt at chunk reading when
catching http.client.IncompleteRead was that once the exception was
thrown, the remaining bytes were no longer available.  Thus, we'd get a
partial read but the end of the response was lost.

This was the reason for the occasional bad json exception.

This patch reworks this with a complete retry strategy now and it works.
Sometimes it takes longer than desired, but it will retry up to 5 times
with increasing timeout.

Signed-off-by: Alex Tsariounov <alext@manetu.com>